Extended logical rewriting to source/target, SetOfSingleton's 'extra' is now a Singleton - Source and target of kernel are also logically rewritten - Extras are checked of type array so as to append information not replace - Name in parmenides is checked non-lemmatized as well as lemmatized - Sped up benchmarking by calling function within single instance rather than calling script itself multiple times - Ensured that min/max values are chosen correctly - 'extra' from SetOfSingletons is now a Singleton, not a string to preserve more information - Added 'INSTRUMENT' rule to logical rewriting   - Added "means verbs" - Catabolites are only deleted for benchmarked sentences when benchmarking - Created assertions for pre-rewriting, original paper, and new

+++ LaSSI/tests/delete_catabolites.py
@@ -4,20 +4,21 @@ from pathlib import Path
 
 
 # Only delete GSM information as no reason to re-get meuDB
-def delete_files(delete_all_files=False):
+def delete_files(delete_all_files=False, benchmarking=False):
     catabolites_dir = os.path.join(Path(os.path.dirname(os.path.abspath(__file__))).parent.absolute().parent.absolute(), "catabolites")
     for subdir, dirs, files in os.walk(catabolites_dir):
-        for dir in dirs:
-            if dir == "viz":
-                dir_path = os.path.join(subdir, dir)
-                print(f"Deleting folder: {dir_path}")
-                shutil.rmtree(dir_path)
-        for file in files:
-            if file in ("gsmDB.txt", "datagramdb_output.json") or (
-                    file in ("internals.json", "internals-bin.json", "string_rep.txt", "meuDBs.json") and delete_all_files):
-                file_path = os.path.join(subdir, file)
-                print(f"Deleting file: {file_path}")
-                os.remove(file_path)
+        if subdir.split('/')[-1][0].isdigit() or not benchmarking:
+            for dir in dirs:
+                if dir == "viz":
+                    dir_path = os.path.join(subdir, dir)
+                    print(f"Deleting folder: {dir_path}")
+                    shutil.rmtree(dir_path)
+            for file in files:
+                if file in ("gsmDB.txt", "datagramdb_output.json") or (
+                        file in ("internals.json", "internals-bin.json", "string_rep.txt", "meuDBs.json") and delete_all_files):
+                    file_path = os.path.join(subdir, file)
+                    print(f"Deleting file: {file_path}")
+                    os.remove(file_path)
 
 
 if __name__ == '__main__':
